pmbus: Update bus type in createPMBus() to uint16_t
Change the bus parameter of PMBus::createPMBus() from uint8_t to
uint16_t. Update all phosphor-power-supply calls accordingly. This
change required to support multi-chassis system, where the bus
identifier may exceed the range of uint8_t.

Tested:
    Simulated the bus ID greater than 255 and verified the apps works as
    expected.

+++ phosphor-power-supply/new_power_supply.cpp
@@ -25,7 +25,7 @@ constexpr auto bindDelay = 1000;
 using namespace sdbusplus::xyz::openbmc_project::Common::Device::Error;
 
 PowerSupply::PowerSupply(
-    sdbusplus::bus_t& bus, const std::string& invpath, std::uint8_t i2cbus,
+    sdbusplus::bus_t& bus, const std::string& invpath, std::uint16_t i2cbus,
     std::uint16_t i2caddr, const std::string& driver,
     const std::string& gpioLineName, std::function<bool()>&& callback,
     const std::string& chassisShortName) :
@@ -102,7 +102,7 @@ PowerSupply::PowerSupply(
 }
 
 PowerSupply::PowerSupply(
-    sdbusplus::bus_t& bus, const std::string& invpath, std::uint8_t i2cbus,
+    sdbusplus::bus_t& bus, const std::string& invpath, std::uint16_t i2cbus,
     std::uint16_t i2caddr, const std::string& driver,
     const std::string& gpioLineName, std::function<bool()>&& callback) :
     PowerSupply(bus, invpath, i2cbus, i2caddr, driver, gpioLineName,
