+++ pmbus.cpp
@@ -404,7 +404,7 @@ void PMBus::findHwmonDir()
     }
 }
 
-std::unique_ptr<PMBusBase> PMBus::createPMBus(std::uint8_t bus,
+std::unique_ptr<PMBusBase> PMBus::createPMBus(std::uint16_t bus,
                                               const std::string& address)
 {
     const std::string physpath = {
@@ -414,7 +414,7 @@ std::unique_ptr<PMBusBase> PMBus::createPMBus(std::uint8_t bus,
     return interface;
 }
 
-std::unique_ptr<PMBusBase> createPMBus(std::uint8_t bus,
+std::unique_ptr<PMBusBase> createPMBus(std::uint16_t bus,
                                        const std::string& address)
 {
     return PMBus::createPMBus(bus, address);
